refactor(desktop): move context, workspace, settings to feature dirs
