+++ apps/desktop/node_modules
@@ -1,0 +1,1 @@
+/Users/akhayam/nerd/kay-am/apps/desktop/node_modules
