+++ apps/desktop/src/App.tsx
@@ -4,12 +4,12 @@ import type { SessionId } from '@kay-am/types';
 import { CommandPalette } from './components/CommandPalette';
 import { BootSplash } from './components/BootSplash';
 import { ChatView } from './components/chat/ChatView';
-import { ContextPanel } from './components/ContextPanel';
+import { ContextPanel } from './features/context/components/ContextPanel';
 import { EndSessionDialog } from './components/EndSessionDialog';
-import { SettingsDialog } from './components/SettingsDialog';
-import { ShortcutHelpDialog } from './components/ShortcutHelpDialog';
+import { SettingsDialog } from './features/settings/components/SettingsDialog';
+import { ShortcutHelpDialog } from './features/settings/components/ShortcutHelpDialog';
 import { ToastProvider } from './components/Toast';
-import { WorkspacesSidebar } from './components/WorkspacesSidebar';
+import { WorkspacesSidebar } from './features/workspace/components/WorkspacesSidebar';
 import { useKeyboardShortcut } from './shared/hooks/use-keyboard-shortcut';
 import {
   useAppStore,
