+++ apps/desktop/src/__tests__/a11y/smoke.test.tsx
@@ -88,8 +88,8 @@ import type { WorkspaceId } from '@kay-am/types';
 import { runA11yCheck } from './utils';
 import { NotificationCenter } from '../../features/notifications/components/NotificationCenter';
 import { BootSplash } from '../../components/BootSplash';
-import { SettingsDialog } from '../../components/SettingsDialog';
-import { WorkspacesSidebar } from '../../components/WorkspacesSidebar';
+import { SettingsDialog } from '../../features/settings/components/SettingsDialog';
+import { WorkspacesSidebar } from '../../features/workspace/components/WorkspacesSidebar';
 import { StatusBar } from '../../components/StatusBar';
 import { SkillsPanel } from '../../features/skills/components/SkillsPanel';
 import { SlashCommandPopover } from '../../components/chat/SlashCommandPopover';
