+++ apps/desktop/src/__tests__/context-panel-rail.test.tsx
@@ -63,7 +63,7 @@ import { afterEach, describe, expect, it, vi } from 'vitest';
 import { cleanup, fireEvent, render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import type { Session, SessionId, WorkspaceId } from '@kay-am/types';
-import { ContextPanel } from '../components/ContextPanel';
+import { ContextPanel } from '../features/context/components/ContextPanel';
 
 afterEach(cleanup);
