+++ apps/desktop/src/features/context/components/ContextPanel/index.tsx
@@ -35,7 +35,7 @@ import type {
   TelemetryRecord,
   PullRequestStateKind,
 } from '@kay-am/types';
-import { PlansModal } from '../features/plans/components/PlansModal';
+import { PlansModal } from '../../../../features/plans/components/PlansModal';
 import {
   EMPTY_ARRAY,
   useAppStore,
@@ -46,18 +46,18 @@ import {
   useSessionSlots,
   useSlotHistory,
   useSummarizerStatus,
-} from '../store';
-import { openUrl } from '../shared/lib/editor';
-import { formatError } from '../shared/lib/errors';
-import { tauriGhRunner } from '../features/github/github';
-import { DiffViewerDialog } from '../features/permissions/components/DiffViewerDialog';
-import { GithubCard } from '../features/github/components/Card';
-import { worktreeStatus } from '../features/worktree/worktree';
+} from '../../../../store';
+import { openUrl } from '../../../../shared/lib/editor';
+import { formatError } from '../../../../shared/lib/errors';
+import { tauriGhRunner } from '../../../../features/github/github';
+import { DiffViewerDialog } from '../../../../features/permissions/components/DiffViewerDialog';
+import { GithubCard } from '../../../../features/github/components/Card';
+import { worktreeStatus } from '../../../../features/worktree/worktree';
 import type { WorktreeStatus } from '@kay-am/types';
 import {
   buildCommentAgentArgs,
   buildReviewChangesAgentArgs,
-} from '../features/chat/spawn-from-comment';
+} from '../../../../features/chat/spawn-from-comment';
 
 interface ContextPanelProps {
   session: Session;
