+++ apps/desktop/src/features/settings/components/ShortcutHelpDialog/index.tsx
